# oyo-web/openapi.yaml
openapi: 3.0.3
info:
  title: OYO LTC API
  description: |
    HTTP surface of the `oyo-web` service — wallet management, queue,
    chain inspection, regtest controls. All endpoints are mounted under
    `/api`.

    See `README-OYO.md` at the repository root for the architectural
    background (wallet kinds, send paradigm, confirm-token flow,
    queue semantics).
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8880
    description: prod (systemd --user oyo-web.service)
  - url: http://localhost:8881
    description: test (oyo-dev.sh)

# All endpoints under /api/* are gated by HTTP Basic Auth when the
# server is configured with `webuser` / `webpassword`. Browser users
# get cookie sessions via POST /login but those aren't used by API
# clients — Basic is stateless and works from curl / Python / codegen.
security:
  - basicAuth: []
...
paths:
...
  /api/queue:
    get:
      tags: [queue]
      summary: List queue tasks (optionally filter by wallet)
      parameters:
        - { in: query, name: wallet, schema: { type: string } }
      responses:
        "200":
          description: list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/QueueTask" }
    post:
      tags: [queue]
      summary: Add a task (recipe, optionally scheduled)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QueueTaskBody" }
      responses:
        "200":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueueTask" }
        "400": { $ref: "#/components/responses/Error400" }
        "404": { description: "wallet not found" }
  "/api/queue/{id}":
    parameters:
      - { in: path, name: id, required: true, schema: { type: string }, description: "Task id (16 hex chars)" }
    get:
      tags: [queue]
      summary: Single task detail
      responses:
        "200":
          description: task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueueTask" }
        "404": { description: "task not found" }
    delete:
      tags: [queue]
      summary: Remove a task (frees nothing — recipe storage only)
      responses:
        "200": { description: removed }
        "404": { description: "task not found" }
  "/api/queue/{id}/run":
    post:
      tags: [queue]
      summary: Execute a pending task immediately
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: post-run state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueueTask" }
        "400": { description: "task is not pending" }

  # ───────────────────── explorer ─────────────────────
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        HTTP Basic Auth — every request carries `Authorization: Basic
        base64(user:pass)`. Validated against the server's configured
        `webuser` / `webpassword`.
  responses:
    Error400:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    QueueTaskBody:
      description: |
        Body of POST /api/queue. Mirrors QueueTask recipe fields plus
        an optional `scheduled_at` (unix seconds or RFC3339).
      type: object
      required: [wallet]
      properties:
        wallet:               { type: string }
        to:                   { type: string }
        amount_sat:           { type: integer }
        outputs:
          type: array
          items:
            type: object
            required: [address, amount_sat]
            properties:
              address:    { type: string }
              amount_sat: { type: integer }
        send_all:             { type: boolean }
        fee_rate_sat_per_vb:  { type: integer }
        inputs:
          type: array
          items:
            type: object
            properties:
              txid:       { type: string }
              vout:       { type: integer }
              commitment: { type: string }
        scheduled_at:
          type: string
          description: |
            Optional. Unix seconds (e.g. "1799999999") or RFC3339
            ("2026-05-05T18:00:00Z"). Omit for manual-only
            execution.
    QueueTask:
      type: object
      properties:
        id:                   { type: string, description: "16 hex chars" }
        wallet:               { type: string }
        to:                   { type: string }
        amount_sat:           { type: integer }
        outputs:
          type: array
          items:
            type: object
            properties:
              address:    { type: string }
              amount_sat: { type: integer }
        send_all:             { type: boolean }
        fee_rate_sat_per_vb:  { type: integer }
        inputs:
          type: array
          items:
            type: object
            properties:
              txid:       { type: string }
              vout:       { type: integer }
              commitment: { type: string }
        scheduled_at:         { type: string, format: date-time, nullable: true }
        status:               { type: string, enum: [pending, running, done, failed] }
        txid:                 { type: string, description: "populated on done" }
        error:                { type: string, description: "populated on failed" }
        created_at:           { type: string, format: date-time }
        executed_at:          { type: string, format: date-time, nullable: true }